requirements.txt
flask
python-dotenv
langchain 
langchain-community 
langchain-openai
langchain-mongodb
pymongo
scrapingant-client
chardet